+++ packages/react/src/themes/teams/components/Video/videoStyles.ts
@@ -1,0 +1,13 @@
+import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
+import { VideoProps } from '../../../../components/Video/Video'
+import { VideoVariables } from './videoVariables'
+
+export default {
+  root: ({ variables: v }): ICSSInJSStyle => ({
+    boxSizing: 'border-box',
+    display: 'inline-block',
+    verticalAlign: 'middle',
+    width: v.width,
+    height: v.height || 'auto',
+  }),
+} as ComponentSlotStylesInput<VideoProps, VideoVariables>