+++ CHANGELOG.md
@@ -17,7 +17,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixes
+- Fix overflowing focus outline for `Grid` items for Teams theme @Bugaa92 ([#1195](https://github.com/stardust-ui/react/pull/1195))
+
 ### Features
+- Add `Embed` and `Video` components @stuartlong ([#1108](https://github.com/stardust-ui/react/pull/1108))
 - Add `Chat` + `Reaction` screener tests for dark and hoc Teams themes @mnajdova ([#1204](https://github.com/stardust-ui/react/pull/1204))
 
 <!--------------------------------[ v0.27.0 ]------------------------------- -->